% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_tables.R
\name{extractor_tables}
\alias{extractor_tables}
\title{Extract tables from pdf documents}
\usage{
extractor_tables(
  x,
  path = FALSE,
  rec = FALSE,
  onecol = FALSE,
  delimiter = "\\\\s{2,}",
  delimiter_table = "\\\\s{2,}",
  replacement = "|"
)
}
\arguments{
\item{x}{Either the text of the pdf read in with the pdftools package or a
path for the location of the pdf file.}

\item{path}{TRUE/FALSE; An optional path designation for the location of the pdf to be
converted to text. The pdftools package is used for this conversion.}

\item{rec}{TRUE/FALSE indicating whether the table in the pdf or the page has
a rectangular dimension.That is, all rows and all columns are of equal length
this is important because reading a table without proper dimension will
through an error.}

\item{onecol}{TRUE/FALSE indicating whether the pdf file is one column, used
only by the extract_pages helper function}

\item{delimiter}{A delimiter used to detect tables. The default is two
consecutive blank white spaces.}

\item{delimiter_table}{A delimiter used to separate table cells. The default
value is two consecutive blank white spaces.}

\item{replacement}{A delimiter used to separate table cells after the
replacement of white space is done.}
}
\value{
A list containing data.frame for all extracted table
}
\description{
Using the pdftools package, tables in the pdf document are extracted
as a dataframe and returned as a list.
}
