% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages_tables.R
\name{pages_tables}
\alias{pages_tables}
\title{Extract tables from pdf documents}
\usage{
pages_tables(
  x,
  pages,
  path = FALSE,
  rec = FALSE,
  onecol = FALSE,
  delimiter = "\\\\s{2,}",
  delimiter_table = "\\\\s{2,}",
  replacement = "|"
)
}
\arguments{
\item{x}{Either the text of the pdf read in with the pdftools package or a
path for the location of the pdf file.}

\item{pages}{numeric vector of pages from which table will be extracted.
it can be a single number or a vector of numbers.}

\item{path}{An optional path designation for the location of the pdf to be
converted to text. The pdftools package is used for this conversion.}

\item{rec}{TRUE/FALSE indicating whether the table in the pdf or the page has
a rectangular dimension.That is, all rows and all columns are of equal length
this is important because reading a table without proper dimension will
through an error.}

\item{onecol}{TRUE/FALSE indicating whether the pdf file is one column}

\item{delimiter}{A delimiter used to detect tables. The default is two
consecutive blank white spaces.}

\item{delimiter_table}{A delimiter used to separate table cells. The default
value is two consecutive blank white spaces.}

\item{replacement}{A delimiter used to separate table cells after the
replacement of white space is done.}
}
\value{
A list containing data.frame for all extracted table
}
\description{
Using the pdftools package, tables in the pdf document are extracted
as a dataframe and returned as a list.
}
\examples{
file <- system.file("extdata", "onecoldata.pdf", package = "pdftableExtractor")

table2 <- pages_tables(file,pages = 19, path = TRUE)
table2

# extract rectangular data
pages_tables(file, pages = 19, path = TRUE, rec = TRUE)

}
